SEQUOIADBMAINSTREAM-8761 fix compile warning

+++ SequoiaDB/engine/msg/msgMessage.cpp
@@ -682,13 +682,13 @@ INT32 msgExtractInsert ( const CHAR *pBuffer, INT32 *pflag,
          rc = SDB_INVALIDARG ;
          goto error ;
       }
-      if ( (INT32)pInsert->header.messageLength <= offset )
+      if ( pInsert->header.messageLength <= offset )
       {
          break ;
       }
       pCurrent = &pBuffer[offset] ;
    }
-   if ( ossAlign4( pInsert->header.messageLength ) < offset )
+   if ( (INT32)ossAlign4( pInsert->header.messageLength ) < offset )
    {
       rc = SDB_INVALIDARG ;
       goto error ;
